requirements.txt
python-dotenv
discord.py
alpha_vantage
pandas